Add failure callback, remove retries and reorganize module
